% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplots_comparison.R
\name{barplots_comparison}
\alias{barplots_comparison}
\title{Plot barplots for comparison between the conditions at one or more time points}
\usage{
barplots_comparison(
  data,
  time = "default",
  axisTitle = "\% Labeling",
  plotTitle = "Bin"
)
}
\arguments{
\item{data}{object (1) The 'average_labeling' or the 'mol_equivalent_labeling' object created using \link{get_table_objects_NA_corrected} function}

\item{time}{numeric vector, time point(s) of interest. Default is the last time point}

\item{axisTitle}{string (1) y-axis text on the plot}

\item{plotTitle}{string (1) title to be used for barplots i.e. "Bin" or "Compound" column}
}
\value{
The function returns a list that contains all figures generated and if a single time point is selected, a dataframe with compounds that are
labeled significantly different. Plots (and a table) will also be saved to the current directory.
}
\description{
This function takes input time point(s) and extract the label enrichment information for a side-by-side barplot comparison plotting.
When only a single time is supplied, the function also returns a table containing Bins that have significant different labeling between the two
categories at that time point.
}
\examples{
\donttest{try(barplots_comp <- barplots_comparison(test_13C15NGlutamine$average_labelingNAcorrected, time = "default", axisTitle = "\% Labeling", plotTitle = "Compound"))}
}
\seealso{
\link{get_table_objects_NA_corrected}, \link{label_enrichment_plot}
}
