% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Pull out the first field which correpsonds to timepoint of non-stationary labeling experiment}
\usage{
data_clean(x)
}
\description{
Pull out the first field which correpsonds to timepoint of non-stationary labeling experiment
}
