% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanDot.R
\name{data_cleanDot}
\alias{data_cleanDot}
\title{Extract the first field of labels}
\usage{
data_cleanDot(x)
}
\arguments{
\item{x}{string (1) Input string for first field extraction}
}
\value{
Returns the first field from a string separated by "."
}
\description{
Pull out the first field of string separated by dots ('.')
}
