% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanDotII.R
\name{data_cleanDotII}
\alias{data_cleanDotII}
\title{Extract the second field of labels}
\usage{
data_cleanDotII(x)
}
\arguments{
\item{x}{string (1) Input string for second field extraction}
}
\value{
Returns the second field from a string separated by "."
}
\description{
Pull out the second field of character separated dots ('.')
}
