% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanII.R
\name{data_cleanII}
\alias{data_cleanII}
\title{Pull out the second field which correpsonds to categroy of sample in the non-stationary labeling experiment}
\usage{
data_cleanII(x)
}
\description{
Pull out the second field which correpsonds to categroy of sample in the non-stationary labeling experiment
}
\examples{
data_cleanII()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
