% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanIII.R
\name{data_cleanIII}
\alias{data_cleanIII}
\title{Extract the third field of the column labels}
\usage{
data_cleanIII(x)
}
\arguments{
\item{x}{string (1) Input string for third field extraction}
}
\value{
Returns the third field from a string separated by "\_"
}
\description{
Pull out the third field which corresponds to replicate of sample of non-stationary labeling experiment
}
