% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_mass.R
\name{get_comp_mass}
\alias{get_comp_mass}
\title{This function is calculating the m/z for the chemical formulae provided}
\usage{
get_comp_mass(comp_formula, polarity)
}
\arguments{
\item{comp_formula}{is the formula from the annotation file}

\item{polarity}{will be the polarity from both Compund_Data and XCMS_data}
}
\description{
This function is calculating the m/z for the chemical formulae provided
}
\examples{
get_comp_mass()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
