% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_mz_lookup.R
\name{get_comp_mz_lookup}
\alias{get_comp_mz_lookup}
\title{Extract information on isotopologues to be searched for}
\usage{
get_comp_mz_lookup(
  compound_data,
  comp_formula,
  r_time,
  ppm,
  polarity,
  label_scheme
)
}
\arguments{
\item{compound_data}{is the annotation file}

\item{comp_formula}{is the formula}

\item{r_time}{This will be the rt from your annotation file}

\item{ppm}{is the ppm error to be used to get a range of possible mass}

\item{polarity}{will be the polarity from both Compound_Data and XCMS_data}

\item{label_scheme}{defines the labeling scheme of the experiments}
}
\value{
The function returns a list of various information of possible isotopologues including:
mass, lower bound and upper bound calculated based on ppm error provided, as well as numbers of
isotopes
}
\description{
This function is going to evaluate chemical formulas to determine a set of possible isotopologues
}
