% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_stage.R
\name{get_comp_stage}
\alias{get_comp_stage}
\title{This function is going to evaluate all of the XCMS_data to identify isotopologues}
\usage{
get_comp_stage(x, y, comp_lookup_table, r_time, rt_tolerance)
}
\arguments{
\item{x}{this is the m/z from your XCMS_data}

\item{y}{will be RT from your XCMS data}

\item{comp_lookup_table}{This will be your m/z table that was created based on the annotation file}

\item{r_time}{This will be the rt from your annotation file}

\item{rt_tolerance}{will be the retention time window for the upper and lower bounds}
}
\description{
This function is going to evaluate all of the XCMS_data to identify isotopologues
}
\examples{
get_comp_stage()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
