% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_element_count.R
\name{get_element_count}
\alias{get_element_count}
\title{This function is calculating the number of each of the elements present in the formula}
\usage{
get_element_count(comp_formula)
}
\arguments{
\item{comp_formula}{is the formula from the annotation file}
}
\description{
This function is calculating the number of each of the elements present in the formula
}
\examples{
get_element_count()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
