% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_objects.R
\name{get_table_objects}
\alias{get_table_objects}
\title{Function to process XCMS data for isotopic enrichment and create MID and average labeling objects
from non-stationary isotopic labeling experiments}
\usage{
get_table_objects(
  XCMS_data,
  compounds_data,
  ppm = 10,
  rt_tolerance = 0.1,
  output = NULL
)
}
\arguments{
\item{XCMS_data}{this will be your edited XCMS output}

\item{compounds_data}{This will be your annotation file.}

\item{ppm}{will be ppm for theupper and lower bounds}

\item{rt_tolerance}{will be the retention time window for the upper and lower bounds}

\item{output}{This is the name which will be appended to the files if the user wishes to print out the analysis tables}
}
\description{
Function to process XCMS data for isotopic enrichment and create MID and average labeling objects
from non-stationary isotopic labeling experiments
}
\examples{
get_table_objects()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
