% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_object_NA_corrected.R
\name{get_table_objects_NA_corrected}
\alias{get_table_objects_NA_corrected}
\title{Identifying single and dual labeled isotopologues from XCMS lists for chemical formulas provided.
Create natural abundance corrected post-processing data tables for pre-processed, stable isotope labeling based metabolomics datasets}
\usage{
get_table_objects_NA_corrected(
  XCMS_data,
  compounds_data,
  ppm = 10,
  rt_tolerance = 0.1,
  output = "sample_output",
  label_scheme
)
}
\arguments{
\item{XCMS_data}{XCMS pre-processed data provided as user input in .csv format}

\item{compounds_data}{User annotation file (in .txt format) containing a list of compounds for which isotopologues are to be identified.}

\item{ppm}{numeric (1) define the maximal tolerated deviation in m/z (as parts per million) from calculate labeled isotopologue within XCMS_data}

\item{rt_tolerance}{numeric (1) the maximal tolerate retention time deviation in minutes (from specified value provided in the annotation file) when searching for labeled isotopologues within XCMS_data.}

\item{output}{A title to append to the files if the user wishes to print out the analyses tables/objects.}

\item{label_scheme}{string (1) The labeling scheme of the experiments. It currently supports single and dual labeled settings that involve
13C, 15N, and 2H. All available options are: "C", "N", "H", "CN", "CH", "NH".}
}
\value{
The function returns a class instance with eight objects 1. MIDs 2. scaled_MIDs 3. average_labeling 4. mol_equivalent_labeling as well as natural
abundance corrected versions of 1-4.
In addition, the function also exports the eight objects as .txt files containing the specific data tables for user verification into two
separate directories.
}
\description{
Function to process XCMS data for stable isotope enriched metabolomics experiments. This function first calculates all the
isotopologue m/z for a list of chemical formulae provided by the user as "compound_data" and identifies the matched m/z features within
the "xcms_data". In addition, natural abundance correction is achieved using the package IsoCorrectoR.
Four objects are created using this function, an MIDs table containing mass isotopologue distribution matrix for all the
compounds, a scaled_MIDs table where the MIDs are adjusted by pool sizes provided in the compounds data to account for differences in pool sizes of compounds,
an average_labeling table where the label enrichment for each compound is represented as a percentage and finally a mol_equivalent_labeling
table where the label enrichment is represented as mol equivalents of labeled compound at each time point. Four additional objects similar to the above
but corrected for natural abundance are also created.
}
\examples{
olddir <- setwd(tempdir())
\donttest{try(test_13C15NGlutamine <- get_table_objects_NA_corrected(XCMS_data, compounds_data,
ppm = 5, rt_tolerance = 0.1, label_scheme = "CN", output = "13C15N_Glutamine"))}
setwd(olddir)
}
\seealso{
\link{MIDplot}, \link{label_enrichment_plot}, \link{PCA_and_heatmap}, \link{getClustersAndPlots}
}
