% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabelEnrichment.R
\name{getLabelEnrichment}
\alias{getLabelEnrichment}
\title{Calculate label enrichment}
\usage{
getLabelEnrichment(
  MIDs_table,
  forMedianNormalization,
  columnsToSearch,
  subsetOfTableJustAnnotationData,
  compounds_data
)
}
\arguments{
\item{MIDs_table}{The MIDs table}

\item{forMedianNormalization}{A dataframe used for calculation}

\item{columnsToSearch}{All columns that contain experiment data}

\item{subsetOfTableJustAnnotationData}{A subset of data with annotation information columns}

\item{compounds_data}{The compounds data}
}
\description{
This function is going to calculate the label enrichment and is called in \link{get_table_objects_NA_corrected}
}
