% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_enrichment_plot.R
\name{label_enrichment_plot}
\alias{label_enrichment_plot}
\title{Function to plot time-course label enrichment both as average_labeling and mole_equivalents_labeling
in a non-stationary isotopic labeling experiment}
\usage{
label_enrichment_plot(
  mydata,
  Category,
  yLim = NULL,
  xLim = NULL,
  axisTitle = "Labeling",
  plotTitle = "Bin",
  plotTitle2 = NULL,
  outputName = "_"
)
}
\arguments{
\item{mydata}{object (1) The 'average_labeling' or the 'mol_equivalent_labeling' object created using get_table_objects() function}

\item{Category}{string (1) name of the category by which you are binning, for the example data set - Lipidomics has two categories (i.e. tissue types),
"Cotyledon" and "EA" and the DualLabel dataset has two catergories (i.e. genotypes), "WT" and "GAT"}

\item{axisTitle}{string (1) The label to be used for the Y-Axis. Typically it is "\% labeling" for average_labeling and "mol equivalents of label" for mol_equivalent_labeling}

\item{plotTitle}{string (1) The title to be used for the plot - has to be one of the columns within the object.
Typically, it is "Compound", however, if the string is too long in the 9x9 output generated, "Bin" can be used as a substitute}

\item{plotTitle2}{string (1) Additional title to be used for the plot, default is set to NULL}

\item{outputName}{string (1) The name to be appended to the the output pdf}

\item{ylim}{numeric (1) sets the limit of y-axis, default is set to maximum}

\item{xlim}{numeric (1) sets the limit of the x-axis, default is set to maximum}
}
\value{
The function returns a pdf file with plots for label enrichment for all the compounds identified within the XCMS_data
}
\description{
This functions allows plotting of the average_labeling and mol_equivalent_labeling objects for user interpretation of label enrichment over the course of the experiment
}
\examples{
Average_Labeling <- label_enrichment_plot(test_13C15NGlutamine$average_labeling, Category = "WT", yLim=NULL,xLim=NULL, axisTitle=" \% Labeling", plotTitle2= "Bin", outputName = "Average")
Label_Enrichment <- label_enrichment_plot(test_13C15NGlutamine$mol_equivalent_labeling, Category = "WT", yLim=NULL,xLim=NULL, axisTitle="mol equivalents of Label", plotTitle="Compound", plotTitle2= "Bin", outputName = "mol Equi")
}
\seealso{
PCA_and_heatmap(), MIDplot(), getClustersAndPlots(), get_table_objects()
}
\author{
Shrikaar Kambhampati, Allen Hubbard
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
