% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDs_comparison.R
\name{MIDs_comparison}
\alias{MIDs_comparison}
\title{Plot barplots for MIDs comparison between the conditions at a selected time point}
\usage{
MIDs_comparison(data, time, axisTitle = "MID", plotTitle = "Bin")
}
\arguments{
\item{data}{object (1) the "MIDs" or the "scaled_MIDs" object created using \link{get_table_objects_NA_corrected} function}

\item{time}{numeric (1), time point of interest}

\item{axisTitle}{string (1) y-axis text on the plot}

\item{plotTitle}{string (1) title to be used for barplots i.e. "Bin" or "Compound" column}
}
\value{
The function returns a list that contains all figures generated. Plots will also be saved to the current directory.
}
\description{
This function takes an input time point and extract the MIDs for a barplot comparison plotting.
}
\examples{
\donttest{try(MIDs_comp <- MIDs_comparison(test_13C15NGlutamine$MIDs_tableNAcorrected, time = 8, axisTitle = "MID", plotTitle = "Compound"))}
}
\seealso{
\link{get_table_objects_NA_corrected}, \link{MIDplot}
}
