% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDsTableNoScale.R
\name{MIDsTableNoScale}
\alias{MIDsTableNoScale}
\title{Calculate MIDs table and average labeling table}
\usage{
MIDsTableNoScale(XCMS_data, label_scheme)
}
\arguments{
\item{XCMS_data}{The dataframe after searching for isotopologues}

\item{label_scheme}{The labeling scheme of the experiments}
}
\value{
The function returns a list containing: an MIDs table, an average labeling table and a vector of
bins that do not have an M0
}
\description{
This function is going to calculate the unscaled MIDs Table and is called in get_table_objects_NA_corrected()
}
