% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NACorrectionFxnII.R
\name{NACorrectionFxnII}
\alias{NACorrectionFxnII}
\title{Perform natural abundance correction}
\usage{
NACorrectionFxnII(MIDs_table, label_scheme, output = "_")
}
\arguments{
\item{MIDs_table}{The MIDs table calculated}

\item{label_scheme}{The labeling scheme of the experiments}

\item{output}{The output directory}
}
\value{
The function returns a dataframe of a corrected MIDs table
}
\description{
This function is going to do the NA correction and is called in get_table_objects_NA_corrected()
}
