% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Extract the first field of the column labels}
\usage{
data_clean(x)
}
\arguments{
\item{x}{string (1) Input string for first field extraction}
}
\value{
Returns the first field from a string separated by "\_"
}
\description{
Pull out the first field which corresponds to timepoint of non-stationary labeling experiment
}
