% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleanII.R
\name{data_cleanII}
\alias{data_cleanII}
\title{Extract the second field of the column labels}
\usage{
data_cleanII(x)
}
\arguments{
\item{x}{string (1) Input string for second field extraction}
}
\value{
Returns the second field from a string separated by "\_"
}
\description{
Pull out the second field which corresponds to category of sample of non-stationary labeling experiment
}
