% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_mass.R
\name{get_comp_mass}
\alias{get_comp_mass}
\title{Calculate m/z for given formulas}
\usage{
get_comp_mass(comp_formula, polarity = NULL)
}
\arguments{
\item{comp_formula}{Formula of the compound}

\item{polarity}{Polarity of the compound, can be "pos" or "neg" or \code{NULL}. Default to \code{NULL} for a neutral mass.}
}
\value{
The function returns the m/z value calculated
}
\description{
This function is calculating the m/z for a chemical formula provided
}
