% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_mass.R
\name{get_comp_mass}
\alias{get_comp_mass}
\title{Calculate m/z for given formulas}
\usage{
get_comp_mass(comp_formula, polarity)
}
\arguments{
\item{comp_formula}{is the formula from the annotation file}

\item{polarity}{will be the polarity from both Compound_Data and XCMS_data}
}
\value{
The function returns the m/z value calculated
}
\description{
This function is calculating the m/z for the chemical formulas provided
}
