% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_element_count.R
\name{get_element_count}
\alias{get_element_count}
\title{Calculate number of elements in a given formula}
\usage{
get_element_count(comp_formula)
}
\arguments{
\item{comp_formula}{is the formula from the annotation file}
}
\value{
This function returns a list of counts of each element present in the formula
}
\description{
This function is calculating the number of each of the elements present in the formula
}
\examples{
require(dplyr)
get_element_count("C5H5N5")
}
