% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClustersAndPlots.R
\name{getClusterAndPlots}
\alias{getClusterAndPlots}
\title{Function to do statistical analyses and create plots from non-stationary labeling experiments}
\usage{
getClusterAndPlots(
  mydata1,
  mydata2,
  Category,
  nClust = 0,
  labels = "Bin",
  doMIDs = FALSE,
  outputName = "_"
)
}
\arguments{
\item{mydata1}{will be your average labeling OR your mole equivalent average labeling}

\item{mydata2}{will be your MIDs or scaled MIDs}

\item{Category}{name of the category by which we'll be binning}

\item{nClust}{how many kmeans clusters}

\item{labels}{label datapoints by either the "Bin" or "Compound" column}

\item{doMIDs}{do we want to do the plots for the MIDs as well?}

\item{outputName}{This will be the name that will be appended to the the output pdf}
}
\description{
This function allows you to performs kmeans clustering analyses using label enrichment data
}
\examples{
getClustersAndPlots()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
