% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_enrichment_plot.R
\name{label_enrichment_plot}
\alias{label_enrichment_plot}
\title{Function to plot time-course label enrichment both as average_labeling and mole_equivalents_labeling
in a non-stationary isotopic labeling experiment
XCMS data and start getting isotopic information from the annotation file
and add information columns to be used as input for next steps}
\usage{
label_enrichment_plot(
  mydata,
  Category,
  yLim = NULL,
  xLim = NULL,
  axisTitle = "Labeling",
  plotTitle = "Bin",
  plotTitle2 = NULL,
  outputName = "_"
)
}
\arguments{
\item{mydata}{you can use average_labeling or mole_equivalents_labeling as an input for this function}

\item{Category}{will be the tissue type, i.e. WT/Mut, or treatment, i.e treated vs untreated}

\item{axisTitle}{This will be name for the Y-Axis typically it is "\% label  enrichment" for average_labeling and "mole_equivalents" for mole_equivalents_labeling)}

\item{plotTitle}{This may be any one of the columns used in the get_table_objects object, for example: "Bin","Compound" or "Formula"}

\item{plotTitle2}{This is an additional component for the naming and may be any one of the columns used in the get_table_objects object, for example: "Bin","Compound" or "Formula"}

\item{outputName}{This will be the name that will be appended to the the output pdf}

\item{ylim}{is the limit of y-axis}

\item{xlim}{is the limit of the x-axis}
}
\description{
Function to plot time-course label enrichment both as average_labeling and mole_equivalents_labeling
in a non-stationary isotopic labeling experiment
XCMS data and start getting isotopic information from the annotation file
and add information columns to be used as input for next steps
}
\examples{
label_enrichment_plot()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
