% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDplot.R
\name{MIDplot}
\alias{MIDplot}
\title{Plot Mass Isotopologue Distribution, i.e. MIDs, in a non stationary isotopic
labeling experiment}
\usage{
MIDplot(
  myData,
  Category,
  splitIsotopologue = "default",
  axisTitle = "MID",
  plotTitle = "Bin",
  plotTitle2 = NULL,
  yLimit = "default",
  xLimit = "default",
  outputName = "_"
)
}
\arguments{
\item{myData}{object (1) The "MIDs" or the "scaled_MIDs" object created using \link{get_table_objects_NA_corrected} function}

\item{Category}{string (1) name of the category by which you are binning, for the example data set - Lipidomics has two categories (i.e. tissue types),
"Cotyledon" and "EA" and the DualLabel dataset has two categories (i.e. genotypes), "WT" and "GAT"}

\item{splitIsotopologue}{string (1) The isotopologue to be plotted separately in "split_MIDs". Generally if there is low amount of labeling in certain compounds,
splitting the MID plot by the unlabeled isotopologue (i.e. "C0N0" - set as default) will allow better visualization of label enrichment in the other isotopologues}

\item{axisTitle}{string (1) The label to be used for the Y-Axis. Typically it is "MIDs" for MIDs and "mol equivalents of label" for scaled_MIDs}

\item{plotTitle}{string (1) The title to be used for the plot - has to be one of the columns within the object. Typically, it is "Compound", however, if the string is too long in the 9x9 output generated, "Bin" can be used as a substitute}

\item{plotTitle2}{string (1) Additional title to be used for the plot, default is set to NULL}

\item{yLimit}{numeric (1) sets the limit of y-axis, default is set to maximum}

\item{xLimit}{numeric (1) sets the limit of the x-axis, default is set to maximum}

\item{outputName}{string (1) The name to be appended to the the output pdf}
}
\value{
The function returns all plots generated along with two pdf files one with the MIDs of individual compounds as a single plot and the other with MIDs split by a user specified isotopologue for better visualization
of low labeled isotopologues
}
\description{
This function allows plotting of the MID and scaled_MID objects for user interpretation of isotopologue specific label enrichment over the course of the experiment
}
\examples{
\donttest{try(MIDs <- MIDplot(test_13C15NGlutamine$MIDs, Category = "WT",
splitIsotopologue = "C0N0", axisTitle = "MID", plotTitle = "Bin"))}
\donttest{try(scaled_MIDs <- MIDplot(test_DualLabel$scaled_MIDs, Category = "GAT",
axisTitle = "mol_equivalents", plotTitle = "Compound", plotTitle2 = "Bin", outputName = "scaled"))}
}
\seealso{
\link{get_table_objects_NA_corrected}, \link{label_enrichment_plot}, \link{PCA_and_heatmap}, \link{getClustersAndPlots}
}
