% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDplot.R
\name{MIDplot}
\alias{MIDplot}
\title{Function to plot Mass Isotopologue Distribution, i.e. MIDs, in a non stationary isotopic
labeling experiment}
\usage{
MIDplot(
  myData,
  Category,
  splitIsotopologue = "C0N0",
  axisTitle = "MID",
  plotTitle = "Bin",
  plotTitle2 = NULL,
  yLimit = "default",
  xLimit = "default",
  outputName = "_"
)
}
\arguments{
\item{myData}{you can use MIDs or scaled_MIDs as an input for this function}

\item{Category}{will be the tissue type, i.e. WT/Mut, or treatment, i.e treated vs untreated}

\item{axisTitle}{This will be name for the Y-Axis typically it is "MIDs" for MIDs and "mole equivalents of signals" for scaled_MIDs)}

\item{plotTitle}{This will be name for the Y-Axis typically it is "\% label  enrichment" for average_labeling and "mole_equivalents" for mole_equivalents_labeling)}

\item{plotTitle2}{This will be name for the Y-Axis typically it is "\% label  enrichment" for average_labeling and "mole_equivalents" for mole_equivalents_labeling)}

\item{outputName}{This will be the name that will be appended to the the output pdf}

\item{ylim}{is the limit of y-axis}

\item{xlim}{is the limit of the x-axis}
}
\description{
Function to plot Mass Isotopologue Distribution, i.e. MIDs, in a non stationary isotopic
labeling experiment
}
\examples{
MIDplot()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
