% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_and_heatmap.R
\name{PCA_and_heatmap}
\alias{PCA_and_heatmap}
\title{Perform global analysis of label enrichment and generate PCA plots and heatmaps}
\usage{
PCA_and_heatmap(
  mydata1,
  PCMax = 3,
  heatMapCategories,
  labels = "Bin",
  outputName = "_"
)
}
\arguments{
\item{mydata1}{object (1) The 'average_labeling' or the 'mol_equivalent_labeling' object created using \link{get_table_objects_NA_corrected} function}

\item{PCMax}{numeric (1) maximum numbers of PC's to plot}

\item{heatMapCategories}{string (1) or (2) c("Category1") or c("Category1", "Category2") The category to be used for plotting a heatmap}

\item{labels}{string (1) label to be used for heat maps i.e. "Bin" or "Compound" column}

\item{outputName}{string (1) The name to be appended to the the output pdf}
}
\value{
The function returns a list containing generated figures along with a pdf file with PCA plots comparing all PCs up to the user specified number and heat maps with label enrichment
for all compounds using the user specified Categories
}
\description{
This function performs Principle Component Analysis and plots the average_labeling or mol_equivalent_labeling data in a two dimensional space
to allow comparison of time course labeling enrichment data obtained from a non-stationary isotopic labeling experiment. In addition, this function
also plots the label enrichment for each of the compounds as a heat map to easily identify compounds that have significant labeling using global view
}
\examples{
\donttest{try(PCA_Heatmap_Avg <- PCA_and_heatmap(test_13C15NGlutamine$average_labeling, PCMax = 3,
heatMapCategories = c("WT", "GAT"), labels="Bin", outputName = "Average"))}
\donttest{try(PCA_Heatmap_molEqui <- PCA_and_heatmap(test_13C15NGlutamine$mol_equivalent_labeling, PCMax = 3,
heatMapCategories = c("WT", "GAT"), labels="Bin", outputName = "mol_equivalents"))}
}
\seealso{
\link{get_table_objects_NA_corrected}, \link{MIDplot}, \link{label_enrichment_plot}, \link{getClustersAndPlots}
}
