% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_and_heatmap.R
\name{PCA_and_heatmap}
\alias{PCA_and_heatmap}
\title{Function to plot Mass Isotopologue Distribution, i.e. MIDs, in a non stationary isotopic
labeling experiment}
\usage{
PCA_and_heatmap(mydata1, PCMax = 3, heatMapCategories, labels = "Bin")
}
\arguments{
\item{PCMax}{maximum numbers of PC's to plot}

\item{heatMapCategories}{will be the tissue type(s), i.e. WT/Mut, or treatment, i.e treated vs untreated or both.  A heatmap will be conducted for the data associated with each category}

\item{labels}{label datapoints by either the "Bin" or "Compound" column}

\item{mydata}{you can use average_labeling or mole_equivalents_labeling as an input for this function}
}
\description{
Function to plot Mass Isotopologue Distribution, i.e. MIDs, in a non stationary isotopic
labeling experiment
}
\examples{
PCA_and_heatmap()
}
\keyword{MS1}
\keyword{dual}
\keyword{isotopes,}
\keyword{isotopic}
\keyword{labeling,}
\keyword{labels,}
\keyword{metabolomics,}
\keyword{non-stationary}
\keyword{stable}
\keyword{untargeted}
